move a type declaration to its home
