+++ app/types/recommendation.ts
@@ -1,4 +1,6 @@
-
+/**
+ * A crash recommendation object for the fatality review board component
+ */
 export type Recommendation = {
   id: number;
   created_at: string;
@@ -11,11 +13,19 @@ export type Recommendation = {
   recommendations_partners: RecommendationPartner[] | null;
 };
 
+/**
+ * Recomendation status lookup item
+ */
 export type RecommendationStatus = {
   id: number;
   rec_status_desc: string;
 };
 
+/**
+ * A recommendation partner object which enables the
+ * many to many relationship between recommendations and
+ * recommendation partners
+ */
 export type RecommendationPartner = {
   id: number;
   partner_id: number | null;
@@ -23,7 +33,28 @@ export type RecommendationPartner = {
   atd__coordination_partners_lkp: CoordinationPartner | null;
 };
 
+/**
+ * A coordination partner lookup item
+ */
 export type CoordinationPartner = {
   id: number;
   coord_partner_desc: string;
+};
+
+/**
+ * A partial recommendation type which is more flexible
+ * so that it can be used with edit components
+ */
+export type RecommendationFormInputs = {
+  id?: number | undefined;
+  created_at?: string | undefined;
+  created_by?: string | undefined;
+  crash_pk?: number | undefined;
+  rec_text?: string | null | undefined;
+  rec_update?: string | null | undefined;
+  recommendation_status_id?: number | null | undefined;
+  recommendations_partners?:
+    | Partial<RecommendationPartner>[]
+    | null
+    | undefined;
 };
